Thread utility. Added event thread
